Loose ends before Files
I have finished adding drag n color functionality with the LMB, and erasing capability with the RMB. As well as the magnification dropdown in the preview window. Then finally added the functionality to the text box in the color selection window, so when it's changed, the current color, color wheel, and color preview square adjusts to the new color.